refactor: Move things out of client
